merged

+++ plugins/webgui_order/index.js
@@ -27,15 +27,15 @@ const getOrdersAndAccountNumber = async () => {
     'webgui_order.autoRemove as autoRemove',
     knex.raw('count(account_plugin.id) as accountNumber'),
   ])
-  .leftJoin('account_plugin', 'account_plugin.orderId', 'webgui_order.id')
-  .groupBy('webgui_order.id')
-  .orderBy('webgui_order.name', 'ASC');
+    .leftJoin('account_plugin', 'account_plugin.orderId', 'webgui_order.id')
+    .groupBy('webgui_order.id')
+    .orderBy('webgui_order.name', 'ASC');
   return orders;
 };
 
 const getOneOrder = async orderId => {
   const order = await knex('webgui_order').where({ id: orderId }).then(s => s[0]);
-  if(!order) { return Promise.reject('order not found'); }
+  if (!order) { return Promise.reject('order not found'); }
   return order;
 };
 
@@ -44,12 +44,12 @@ const getOneOrderByAccountId = async accountId => {
     'webgui_order.id as id',
     'webgui_order.changeOrderType as changeOrderType',
   ]).leftJoin('account_plugin', 'account_plugin.orderId', 'webgui_order.id')
-  .where({ 'account_plugin.id': accountId }).then(s => s[0]);
+    .where({ 'account_plugin.id': accountId }).then(s => s[0]);
   return order;
 };
 
 const newOrder = async data => {
-  await knex('webgui_order').insert({
+  const [id] = await knex('webgui_order').insert({
     baseId: data.baseId,
     name: data.name,
     shortComment: data.shortComment,
@@ -67,8 +67,9 @@ const newOrder = async data => {
     multiServerFlow: data.multiServerFlow,
     changeOrderType: data.changeOrderType,
     active: data.active,
+    connector: data.connector
   });
-  return;
+  return id;
 };
 
 const editOrder = async data => {
@@ -90,6 +91,7 @@ const editOrder = async data => {
     multiServerFlow: data.multiServerFlow,
     changeOrderType: data.changeOrderType,
     active: data.active,
+    connector: data.connector
   }).where({
     id: data.id,
   });
@@ -97,12 +99,19 @@ const editOrder = async data => {
 };
 
 const deleteOrder = async orderId => {
-  const hasAccount = await knex('account_plugin').where({ orderId });
-  if(hasAccount.length) { return Promise.reject('account with this order exists'); }
-  const isGiftCardOn = config.plugins.giftcard && config.plugins.giftcard.use;
-  const hasGiftcard = isGiftCardOn ? await knex('giftcard').where({ orderType: orderId, status: 'AVAILABLE' }) : [];
-  if(hasGiftcard.length) { return Promise.reject('giftcard with this order exists'); }
-  await knex('webgui_order').delete().where({ id: orderId });
+  const orderInfo = await knex('webgui_order').where({ id: orderId }).then(s => s[0]);
+  if (orderInfo.baseId) {
+    await knex('webgui_order').delete().where({ id: orderId });
+  } else {
+    const hasAccount = await knex('account_plugin').where({ orderId });
+    if (hasAccount.length) { return Promise.reject('account with this order exists'); }
+    const isGiftCardOn = config.plugins.giftcard && config.plugins.giftcard.use;
+    const hasGiftcard = isGiftCardOn ? await knex('giftcard').where({ orderType: orderId, status: 'AVAILABLE' }) : [];
+    if (hasGiftcard.length) { return Promise.reject('giftcard with this order exists'); }
+    const hasFlowPackOrder = await knex('webgui_order').where({ baseId: orderId });
+    if (hasFlowPackOrder.length) { return Promise.reject('flowpack order exists'); }
+    await knex('webgui_order').delete().where({ id: orderId });
+  }
   return;
 };
